整合Servlet,优化代码
