--- src/servlet/HelloServlet.java
@@ -1,38 +1,0 @@
-package servlet;
-
-import bean.User;
-import utils.DBUtils;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.LinkedList;
-import java.util.logging.Logger;
-
-/**
- * Created by  waiter on 18-7-6.
- *
- * @author waiter
- */
-@WebServlet(name = "HelloServlet",urlPatterns = "/hello")
-public class HelloServlet extends HttpServlet {
-    Logger logger=Logger.getLogger(this.getClass().getName());
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-
-        int i=DBUtils.getObjectCount(User.class);
-        logger.info(""+i);
-
-
-
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-
-    }
-}
