+++ src/servlet/LoginServlet.java
@@ -17,14 +17,37 @@ import java.util.logging.Logger;
  *
  * @author waiter
  */
-@WebServlet(name = "LoginServlet",urlPatterns = "/do_login")
+@WebServlet(name = "LoginServlet",urlPatterns = {"/do_login","/do_logout"})
 public class LoginServlet extends HttpServlet {
     UserService userService=UserService.getUserService();
     Logger logger=Logger.getLogger(this.getClass().getName());
 
 
     @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        String uri = request.getRequestURI();
+        String substring = uri.substring(1, uri.length());
+        if("do_logout".equals(substring)){
+            doLogout(request,response);
+        }else {
+            request.getRequestDispatcher("/login.jsp").forward(request,response);
+        }
+    }
+
+    @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        String uri = request.getRequestURI();
+        String substring = uri.substring(1, uri.length());
+        if("do_login".equals(substring)){
+            doLogin(request,response);
+        }else if("do_logout".equals(substring)){
+            doLogout(request,response);
+        }else {
+            request.getRequestDispatcher("/login.jsp").forward(request,response);
+        }
+    }
+
+    protected void doLogin(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
         String username = request.getParameter("username");
         logger.info(request.getRemoteAddr()+"尝试使用用户"+username+"登录");
         User byUserName = userService.findByUserName(username);
@@ -34,8 +57,8 @@ public class LoginServlet extends HttpServlet {
             request.getRequestDispatcher("/login.jsp").forward(request,response);
         }else if (BCrypt.checkpw(request.getParameter("password"),byUserName.getPassWd())){
             logger.info(request.getRemoteAddr()+"尝试使用用户"+username+"登录成功");
-        request.getSession().setAttribute("loginInfo",byUserName);
-        response.sendRedirect("/index.jsp");
+            request.getSession().setAttribute("loginInfo",byUserName);
+            response.sendRedirect("/index.jsp");
         }else {
             request.setAttribute("msg","用户名或密码错误");
             logger.info(request.getRemoteAddr()+"尝试使用用户"+username+"登录失败");
@@ -43,8 +66,12 @@ public class LoginServlet extends HttpServlet {
         }
     }
 
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    protected void doLogout(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
+        Object loginInfo = request.getAttribute("loginInfo");
+        logger.info("用户："+loginInfo+"已退出");
+        request.setAttribute("loginInfo",null);
+        request.setAttribute("msg","退出成功");
         request.getRequestDispatcher("/login.jsp").forward(request,response);
     }
+
 }
